separate Models to own project
