--- DraftDBDiagram/Adress.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace DraftDBDiagram
-{
-    public class UserProfile
-    {
-        public ApplicationUser User { get; set; }
-        public Contact Contact { get; set; }
-    }
-}
