--- DraftDBDiagram/ClassDiagram.cd
@@ -1,59 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram MajorVersion="1" MinorVersion="1">
-  <Class Name="DraftDBDiagram.Message">
-    <Position X="3.75" Y="1" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAABAAAAAAAAgAAAAAAAAAAAAAAAAAQAAAAACAAAAAA=</HashCode>
-      <FileName>Message.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="User" />
-    </ShowAsAssociation>
-  </Class>
-  <Class Name="DraftDBDiagram.Chat">
-    <Position X="2.25" Y="3" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAACEAAAAAAAAAAAAAAAAAAAAAAAAQAAAAA=</HashCode>
-      <FileName>Chat.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsCollectionAssociation>
-      <Property Name="Messages" />
-    </ShowAsCollectionAssociation>
-  </Class>
-  <Class Name="DraftDBDiagram.BaseUser">
-    <Position X="6" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAgAAAAAQAAAAAAAAAAAEAAAAABAAAAAA=</HashCode>
-      <FileName>BaseUser.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="Company" />
-    </ShowAsAssociation>
-  </Class>
-  <Class Name="DraftDBDiagram.Company">
-    <Position X="8.5" Y="1" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>Company.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="DraftDBDiagram.ApplicationUser">
-    <Position X="6" Y="2.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAQAAAAAAAAAAAAAGAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
-      <FileName>ApplicationUser.cs</FileName>
-    </TypeIdentifier>
-    <ShowAsAssociation>
-      <Property Name="BaseUser" />
-      <Property Name="Contact" />
-    </ShowAsAssociation>
-  </Class>
-  <Class Name="DraftDBDiagram.Contact">
-    <Position X="8" Y="3" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAgAQAAAgAAAAAAAAAAEAAAAAIAAAAAAAIAIAAAAAA=</HashCode>
-      <FileName>Contact.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Font Name="Segoe UI" Size="9" />
-</ClassDiagram>
