+++ LiveChat/App_Start/Startup.Auth.cs
@@ -5,7 +5,7 @@ using Microsoft.Owin;
 using Microsoft.Owin.Security.Cookies;
 using Microsoft.Owin.Security.Google;
 using Owin;
-using LiveChat.Models;
+using Models.ModelsMVC;
 
 namespace LiveChat
 {
