+++ LiveChat/Controllers/AccountController.cs
@@ -8,7 +8,7 @@ using System.Web.Mvc;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
-using LiveChat.Models;
+using Models.ModelsMVC;
 
 namespace LiveChat.Controllers
 {
