+++ LiveChat/packages.config
@@ -15,9 +15,9 @@
   <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="1.0.0" targetFramework="net461" />
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.2.3" targetFramework="net461" />
   <package id="Microsoft.Net.Compilers" version="1.0.0" targetFramework="net461" developmentDependency="true" />
-  <package id="Microsoft.Owin" version="3.0.1" targetFramework="net461" />
-  <package id="Microsoft.Owin.Host.SystemWeb" version="3.0.1" targetFramework="net461" />
-  <package id="Microsoft.Owin.Security" version="3.0.1" targetFramework="net461" />
+  <package id="Microsoft.Owin" version="4.0.0" targetFramework="net461" />
+  <package id="Microsoft.Owin.Host.SystemWeb" version="4.0.0" targetFramework="net461" />
+  <package id="Microsoft.Owin.Security" version="4.0.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security.Cookies" version="3.0.1" targetFramework="net461" />
   <package id="Microsoft.Owin.Security.Facebook" version="3.0.1" targetFramework="net461" />
   <package id="Microsoft.Owin.Security.Google" version="3.0.1" targetFramework="net461" />
