+++ Models/ChatModels/Company.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace DraftDBDiagram
+namespace Models.ChatModels
 {
     /// <summary>
     /// This entity may be used in payment system
     /// </summary>
     public class Company
     {
+        public string Name { get; set; }
     }
 }
