+++ Models/ClassDiagram1.cd
@@ -1,0 +1,65 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<ClassDiagram MajorVersion="1" MinorVersion="1">
+  <Class Name="Models.ChatModels.BaseUser">
+    <Position X="3" Y="0.75" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAAAAAAAAAAAgAAAAAAAAAAAAAAAAAEAAAAAAAAAAAA=</HashCode>
+      <FileName>ChatModels\BaseUser.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Class Name="Models.ChatModels.Chat">
+    <Position X="0.5" Y="0.75" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAAAAAAAAAACEAAAAAAAAAAAAAAAAAAAAAAAAQAAAAA=</HashCode>
+      <FileName>ChatModels\Chat.cs</FileName>
+    </TypeIdentifier>
+    <ShowAsCollectionAssociation>
+      <Property Name="Messages" />
+    </ShowAsCollectionAssociation>
+  </Class>
+  <Class Name="Models.ChatModels.Company" Collapsed="true">
+    <Position X="7.5" Y="2.5" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAA=</HashCode>
+      <FileName>ChatModels\Company.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Class Name="Models.ChatModels.Contact">
+    <Position X="3" Y="2.75" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAAgAQAAAgAAAAAAAAAAEAAAAAIAAAAAAAIAIAAAAAA=</HashCode>
+      <FileName>ChatModels\Contact.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Class Name="Models.ChatModels.Message">
+    <Position X="0.5" Y="2.5" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAABAAAAAAAAgAAAAAAAAAAAAAAAAAQAAAAACAAAAAA=</HashCode>
+      <FileName>ChatModels\Message.cs</FileName>
+    </TypeIdentifier>
+    <ShowAsAssociation>
+      <Property Name="User" />
+    </ShowAsAssociation>
+  </Class>
+  <Class Name="Models.ChatModels.UserProfile">
+    <Position X="5.5" Y="0.75" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAQBAAAAAAAAAAAACAQAAAAAAAAAAAAAAAAABAAAAAA=</HashCode>
+      <FileName>ChatModels\UserProfile.cs</FileName>
+    </TypeIdentifier>
+    <ShowAsAssociation>
+      <Property Name="Contact" />
+      <Property Name="User" />
+      <Property Name="Company" />
+      <Property Name="BaseUser" />
+    </ShowAsAssociation>
+  </Class>
+  <Class Name="Models.ModelsMVC.ApplicationUser">
+    <Position X="7.75" Y="0.75" Width="1.5" />
+    <TypeIdentifier>
+      <HashCode>AAAAAAAACAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA=</HashCode>
+      <FileName>ModelsMVC\ApplicationUser.cs</FileName>
+    </TypeIdentifier>
+  </Class>
+  <Font Name="Segoe UI" Size="9" />
+</ClassDiagram>
