+++ Models/ModelsMVC/ManageViewModels.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNet.Identity;
 using Microsoft.Owin.Security;
 
-namespace LiveChat.Models
+namespace Models.ModelsMVC
 {
     public class IndexViewModel
     {
