+++ Models/packages.config
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EntityFramework" version="6.1.0" targetFramework="net461" />
+  <package id="Microsoft.AspNet.Identity.Core" version="2.2.1" targetFramework="net461" />
+  <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.1" targetFramework="net461" />
+  <package id="Microsoft.Owin" version="4.0.0" targetFramework="net461" />
+  <package id="Microsoft.Owin.Security" version="4.0.0" targetFramework="net461" />
+  <package id="Owin" version="1.0" targetFramework="net461" />
+</packages>
